refactored frontend signalR factory
